+++ src/main/java/de/mossgrabers/convertwithmoss/format/nki/type/kontakt5/Kontakt5MonolithType.java
@@ -18,6 +18,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
 
+import de.mossgrabers.convertwithmoss.core.IInstrumentSource;
 import de.mossgrabers.convertwithmoss.core.IMultisampleSource;
 import de.mossgrabers.convertwithmoss.core.INotifier;
 import de.mossgrabers.convertwithmoss.core.IPerformanceSource;
@@ -72,6 +73,14 @@ public class Kontakt5MonolithType extends AbstractKontaktType
             final InputStream dataInputStream = new ByteArrayInputStream (mainFile.data);
             return this.kontakt5Type.readNKI (this.sourceFolder, sourceFile, dataInputStream, metadataConfig, createSamples (monolithFiles));
         }
+    }
+
+
+    /** {@inheritDoc} */
+    @Override
+    public void writeNKM (final OutputStream out, final List<String> safeSampleFolderName, final List<IInstrumentSource> instrumentSources, final int sizeOfSamples) throws IOException
+    {
+        // Not supported
     }